Safari-611.1.10/LayoutTests/fast/parser/numeric-entities-expected.txt

PASS testEntity('&#x0;') is "0xFFFD"
PASS testEntity('&#x10000;') is "0xD800, 0xDC00"
PASS testEntity('&#x10FFFF;') is "0xDBFF, 0xDFFF"
PASS testEntity('&#x110000;') is "0xFFFD"
PASS testEntity('&#xFFFFFF;') is "0xFFFD"
PASS testEntity('&#xFFFFFFF;') is "0xFFFD"
PASS testEntity('&#xFFFFFFFF;') is "0xFFFD"
PASS testEntity('&#xFFFFFFFFF;') is "0xFFFD"
PASS testEntity('&#xA0A103A0;') is "0xFFFD"
PASS testEntity('&#x10000;') is "0xD800, 0xDC00"
PASS testEntity('&#x1000010000;') is "0xFFFD"
PASS testEntity('&#1;') is "0x1"
PASS testEntity('&#11;') is "0xB"
PASS testEntity('&#111;') is "0x6F"
PASS testEntity('&#1111;') is "0x457"
PASS testEntity('&#11111;') is "0x2B67"
PASS testEntity('&#111111;') is "0xD82C, 0xDE07"
PASS testEntity('&#1111111;') is "0xDBFD, 0xDC47"
PASS testEntity('&#11111111;') is "0xFFFD"
PASS testEntity('&#111111111;') is "0xFFFD"
PASS testEntity('&#1111111111;') is "0xFFFD"
PASS testEntity('&#11111111111;') is "0xFFFD"
PASS testEntity('&#x1;') is "0x1"
PASS testEntity('&#x9;') is "0x9"
PASS testEntity('&#xA;') is "0xA"
PASS testEntity('&#xB;') is "0xB"
PASS testEntity('&#xC;') is "0xC"
PASS testEntity('&#xD;') is "0xD"
PASS testEntity('&#x7F;') is "0x7F"
PASS testEntity('&#x81;') is "0x81"
PASS testEntity('&#x8D;') is "0x8D"
PASS testEntity('&#x8F;') is "0x8F"
PASS testEntity('&#x90;') is "0x90"
PASS testEntity('&#x9D;') is "0x9D"
PASS testEntity('&#x80;') is "0x20AC"
PASS testEntity('&#x82;') is "0x201A"
PASS testEntity('&#x83;') is "0x192"
PASS testEntity('&#x84;') is "0x201E"
PASS testEntity('&#x85;') is "0x2026"
PASS testEntity('&#x86;') is "0x2020"
PASS testEntity('&#x87;') is "0x2021"
PASS testEntity('&#x88;') is "0x2C6"
PASS testEntity('&#x89;') is "0x2030"
PASS testEntity('&#x8A;') is "0x160"
PASS testEntity('&#x8B;') is "0x2039"
PASS testEntity('&#x8C;') is "0x152"
PASS testEntity('&#x8E;') is "0x17D"
PASS testEntity('&#x91;') is "0x2018"
PASS testEntity('&#x92;') is "0x2019"
PASS testEntity('&#x93;') is "0x201C"
PASS testEntity('&#x94;') is "0x201D"
PASS testEntity('&#x95;') is "0x2022"
PASS testEntity('&#x96;') is "0x2013"
PASS testEntity('&#x97;') is "0x2014"
PASS testEntity('&#x98;') is "0x2DC"
PASS testEntity('&#x99;') is "0x2122"
PASS testEntity('&#x9A;') is "0x161"
PASS testEntity('&#x9B;') is "0x203A"
PASS testEntity('&#x9C;') is "0x153"
PASS testEntity('&#x9E;') is "0x17E"
PASS testEntity('&#x9F;') is "0x178"
PASS successfullyParsed is true

TEST COMPLETE
